Remove 'font-family: Arial/etc' on two selectors
For some reason, thesae two selectors have hardcode Arial/etc as the font for form legends and H3 elements. Therefore while a front-end contribution page or profile will inherit the site fonts, but in these two instances it will use Arial/etc. 

This can be overriden by the user, but given that there is no other use of Arial in CiviCRM, doesn't seem to make much sense. When theming front-end CiviCRM forms/pages these two elements have to be reset with "font-family: inherit". My guess is it's legacy from a much earlier version of the system (but in making this proposal perhaps the real reason will emerge).