Remove the js alert message when jqueryValidation fails on frontend pages
It's really really ugly & as discussed without it
a) extensions can still opt to display the errors themselves and
b) the inline errors still appear.

Note this will have no real affect on existing code - the ugliness of this has been
a blocker to getting into front end validation & at this stage only the
omnipay extension is trying to work in this space.
https://github.com/eileenmcnaughton/nz.co.fuzion.omnipaymultiprocessor/pull/128

I think ideally with Omnipay it DOES makes sense to display messages
near the checkout button - which might be a future step