[Code-quality] remove references to where_tables, where_clause
I've long been of the impression that the where_tables & where_clause in the Smart group & group tables are not
actually used. On  going through I found places where they are saved to but not any where they are  used.

This removes them - next step would be to actually drop the fields but need Jenkins & another person to agree with me
first