Reduce reliance on multiple specific but confusing class variables
https://github.com/civicrm/civicrm-core/pull/14888 seeks to add auto_renew as a parameter on buildMembershipTypeValues
but on digging the results of that are not used much & could be used less.

It's only because we store them far away from where we use them that the relative silliness is kinda
hidden.

This PR removes one of the reliances.

Note that testSubmitRecurCompleteInstant is a good one for stepping through this