Add test cover for Member_BAO_Query auto_renew field
This adds test coverage to demonstrate that the auto_renew field in the Member_BAO_Query::defaultReturnProperties
achieves nothing.

auto_renew is NOT returned for the export - regardless of it's presence. It IS returned for getRows
but tha is calculated there now in the query object.

Once this is merged I will put up the commit to remove auto_renew from defaultReturnProperties