[Import] Minor cleanup on membership import
We have a nasty construct (perhaps due to php4 limitations) where the code iterates the _mapperKeys
array & sets a class property specifying the key it occupies. We later use this to check it occupies a key.

This could be much more readable by just checking the array at the right time.

Also I removed an unused private property & a require_once to api.php which is always included