Add wrapper class for importProcessor
With the export cleanup creating a new sensible class & migrating code to it turned out to be a good approach.

This adds a class that wraps around the Importer Object, taking the mapping_field db format of fields
as a format - this is also where we wound up with the Export - using that format as it is what is saed.

There is a gap as I noted in https://lab.civicrm.org/dev/core/issues/1172 around the schema's adequacy.

That is not such an issue for this - in that we are not dealing with Contribution import and
this is really just exposing the wrapper for the unit tests at this stage - the only change to
'live' code is a little less php v4 support - we have removed a bunch of these from the constructor of the
other objects already with no observed issues