Add ImportProcessor class & unit tests + extend existing cover
This adds the full ImportProcessor class from the WIP pr #15034 - minus the functions that
replace loadSavedMapping) along with tests. It does NOT start to use the new class as yet.

Note that #15068 will need rebasing if this is merged first & vice versa

The goal is to bring it into use along with extending tests but this is 'safe' in that no funcitonal code is changed