Use Processor to  determine the phone_type_id & im_provider_id
What we are doing here is actually very simple - despite the code being kinda insane - we are just getting
the  fields from the database for  each column in the mapping.

The fields are all loaded into the processor already & those functions are tested. This switches to  using them
and to passing around less variables...

Follow ons will do the same for other fields (website_type_id, location_type_id but I will add more tests as I do that