[Ref] Rationalise dedupe code loop.
We have a function 'dedupePair' which is intended to act on a specific pair
and a loop function which iterates them.

The 'pair actions' and the 'looping actions' are currently jumbled together.

This moves the 'pair actions' to the dedupePair function
and keeps the 'looping actions' in the parent looping function.

Athough I left it out of scope for this PR the api that calls this
function should only call the 'dedupePair' function not the multiple
pair wrapper