Further optimize creditnote_id generation
In case the credit notes started at an offset other than 1, avoid
making a number of getcount calls equal to the offset every time
you refund a contribution.

Before: In a DB with 48,000 contributions having a creditnote_id but
whose largest creditnote_id is 48,100, Civi makes a lookup call for
every creditnote_id between CN_48001 and CN_48101 (100 queries) when
refunding a contribution.

After: In a DB with 48,000 contributions having a creditnote_id but
whose largest creditnote_id is 48,100, Civi makes 3 db queries when
refunding a contribution.