(REF) Remove unused function CRM_Core_Block::setTemplateMenuValues
Note that this is a `private` function, so any hypothetical callers must be
in the same class. There don't appear to be any.

Could there be some dynamic method calls?  Well, you'd expect there to be
some tell-tale such as references to `call_*()` or to a string which builds
up that method name from smaller strings ("set" or "MenuValues" or similar).
Nothing comes up.