Consolidate getFormValues on contribution search
This consolidates 2 places where formValues are determined on contribution search. In one place
force is handled in the other retrieving from the form is handled. It seems Ok to handle both in
one place & call that. This might help iron on some glitches Monish is hitting on #15370

We should probably add saved search retrieval in too

To confirm this works use this url
civicrm/contribute/search?reset=1&reset=1&sort_name=p&receive_date_high=20180101&force=1

- it should filter. If you add another criteria to the form afterwards the url criteria should
be present on the form from the url & the new one should also be respected. If you actively remove the url
criteria on the form it is ignored in the search