dev/financial#72 add 'Template' contribution status
This status exists to exclude template contributions from reports
that don't yet know about the is_template field to be added to
civicrm_contribution. This status is not meant to be available to
be set via the UI, just for the process creating a ContributionRecur
record and associated template contribution.