dev/core#1298 Support custom data for emails in Block::create
In order to do this I have switched Block::create to call the v3 email.create api instead of the BAO::create function.

Currently api v3 supports 'custom fields on any entity.' This is handled at the api layer so it doesn't work via the BAO.

It also turns out not to work (yet) on apiv4, or I haven't nailed the syntax. Either way it seems switching to apiv3 is an
appropriate change at this stage