[REF] Remove transaction.
@mattwire proposed moving this in
https://github.com/civicrm/civicrm-core/pull/15446/commits/8f76cfacf156cfbd11e280b9a86926e0a5082c27

However, I believe removing it is the correct action. The usefullness of transactions is to maintain integrity
by rolling back the creation of a bunch of related entities together if one fails.

However in this case it's not rolling back the contribution or participant record or non discount line items so
if it worked (which I doubt) it would reduce rather than increase db integrity