dev/financial#72 Filter contrib status search selects
Replacing pseudoConstant calls with BAO buildOptions in 'search'
context.

This commit is only intended to touch places where we are building
a list of options to offer to the user, not places where we are
displaying results.

In practice this doesn't remove any currently available statuses,
but now we can filter our the proposed 'Template' status.