Add test for participant receipts, super minor consistency fix.
This is mostly adding testing for participant receipts. The one change to core code is to move the assignment
of contactID to the template from a specific path into the shared code. This parameter is not required by
the templates but it is somewhat common for sites to want to have it assigned to templates for
customisation which is presumably why it's there. Since it is there in one path we should either remove or
have it in all & for a field like contactID all makes sense. Tests lock in that var