Fix Event online receipts to display credit card info if available.
We currently have a very complex IF clause which includes some deprecated parameters.

However, we can simplify this to simply display the information if availble. We only collect
credit card information if we require it & hence the presence of it is an adequate indicator
as to whether to display it.

We could check for credit_card_number instead but actually there is precedent in another tpl
for checking credit_card_type & it's pretty neutral as both should be present or missing