Add github stale workflow
I think we should take advantage of githubs ability to mark things as stale. I'm spending far too much
time dealing with PRs that are irrevocably stalled & I think it's not too much to ask people to take
action on PRs every 2-4 months to show they still care & to allow us to prioritise.

I do think we need to figure out how to hone the various messages. This is the default + I updated days-before-close
to be 60 like days before stale.