Improve metadata support for table civicrm_mailing_job in search
This makes the civicrm_mailing_job_start_date field available as
metadata and cleans up the code specifying how
civicrm_mailing_job gets joined in.

It turns out the code design has a pretty sensible methodology for ensuring that required
tables are added but unaware of this the tables
got configured all over the place in the where clause
functions.