Add uniquenames for mailing_name, mailing_job_status
This is in support of https://github.com/civicrm/civicrm-core/pull/15369

Where the names in the search do not match the 'real names' we have 2 techniques

1) where adding a unique name to the schema gives a match we do that - here that is
the case for mailing_name
2) otherwise we ensure the schema has a unique name if appropriate &
rename the field in the upgrade scriptt

In this case when we look at CRM_Mailing_BAO_Query we don't see mailing_name
as present (only mailing_id which should be an entity reference field). However it
is used in the confusingly-not-related Mailing search so lets rename this
one at the same time.