DB Tpl Strings - Remove unused Smarty function crmDBTpl
I've searched `universe` and haven't found any references to this.  This
appears to be the only code-path which consumes information from
`civicrm_persistent`.

I also went back to spot-check a few SVN branches of CiviCRM (v3.2, v3.3).
It appears that this was added in v3.2 *but never used or documented*.