dev/financial#94 fix eroneous payment creation when editing line items
Per https://lab.civicrm.org/dev/financial/issues/94 an eroneous payment transaction is being created when an extra line is
added to an order. The code comments indicate that during the last refactoring the chunk of code
doing this was identified as highly suspect and on re-review of it I'm sure it should go
as we are altering the items purchases, not recording payment in this routine.

I think it dates back to an earlier concept of the flow where altering what was purchased implictly
altered the payments towards that purchase