Replace log_date with created_date & modified_date in advanced search
This is almost the last step in getting rid of an awful lot of technical debt & hard to maintain code around jcalendar & datepicker.

It does change the UI a bit - the code to create the current UI is pretty tough going and from seeing users try
to work with it I don't think all that code is making it more usable. OTOH we can remove a maintainability blockerr
by switching over