[REF] minor extraction.
Honestly how can it be this hard. I want to use the code to calculate the amount for the selected price set options
for tests since they are jamming in random values currently.

But the code is awful :-( I should be able to extract the function that calculates amount but 'selectMembership'
is jammed in the middle because -yeah. Anyway this doesn't really do much to reduce WTF to code ratio
but it pulls out one tiny bit of relevant string wrangling....