[REF] minor code simplification
Switches a call to an old function to an api function & removes post filtering.

Currently the code calls an old function which gets all membership ids & then filters on
statuses. This switches to an api call & passes statuses in as a filter

Strong test cover in testSingleMembershipForTwoRelationships which tests these lines in a variety of ways