dev/core#1383: Fix Re-Installation of Extensions With Logging Enabled
When uninstalling an extension, logging tables associated to custom groups and
fields will not be deleted. On re-installation, addition of custom fields will
cause DB errors to be thrown, as columns existing on logging tables are tried
to be created again (they already exist on logging tables).

Fixed by checking if the column exists on log table before trying to create
it, treating it as a modification of the schema if it exists.