Add resolved to return array for get_conflicts
The goal here is to be able to display what the script will do if a force mode merge is done. The
expection of this is the reason why the 'conflicts' has a key - to allow 'resolved' to schmooze in next to it
in a non-breaking way.

I haven't added in locations yet as I want to move code around before I do that or the spaghetti will
be breeding