[REF] refactor on nasty Dedupe function
I found with other pieces of refactoring static functions that creating a class to support the refactor
made it much easier as that way I could leverage the fact thatt classes have properties and
get away from the crazy param passing that characterizes a nest of static functions.

This adds a class for that purpose and moves a small chunk of code handling into the class.

The goal is to move the handling is done purely for the form back onto the form....