Fix ambiguous column in search with ORDER BY
This fixes an issue where columns whose names are not unique in a
search query cause a DB error when they're used as a sort column.
The issue can be observed in the contribution search when sorting
by contribution status.

The issue is resolved by using the where field of the column spec,
which holds the fully-qualified name of the column.