Rename activity search field from status_id to activity_status_id
Currently this is on the form as 'status_id' but as soon as it's submitted the value is renamed - I think
we might not even need to change the smart group for this one.

The other field that is being renamed is 'priority_id'. Unless we have an appetite to
add a uniqueName for this field (which I kind of feel like not doing more of for some
cycles), I think we could stop renaming it & rely on the default where handling for the
where clause. From what I can see the reason for prefixing is not about
uniqueness but rather about adding a prefix so it gets handled by
Activity_BAO_Query - but in fact Contact_BAO_Query should handle it just fine
based on metadata